Customize orange button component, created buttons.js for all buttons styles
